+++ src/components/App.js
@@ -2,19 +2,20 @@ import React from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import Playground from './Playground';
 import Embedded from './Embedded';
-import DomEvents from './DomEvents';
+import { PreviewEventsProvider } from '../context/PreviewEvents';
 
 function App() {
   return (
     <Router>
-      <Switch>
-        <Route path="/embed/:gistId?/:gistVersion?" component={Embedded} />
-        <Route path="/events" component={DomEvents} />
-        <Route
-          path={['/gist/:gistId/:gistVersion?', '/']}
-          component={Playground}
-        />
-      </Switch>
+      <PreviewEventsProvider>
+        <Switch>
+          <Route path="/embed/:gistId?/:gistVersion?" component={Embedded} />
+          <Route
+            path={['/gist/:gistId/:gistVersion?', '/']}
+            component={Playground}
+          />
+        </Switch>
+      </PreviewEventsProvider>
     </Router>
   );
 }
